Tema 9
Ejercicio tema 9

+++ Tema 9/src/Main.java
@@ -32,10 +32,10 @@ class Persona{
 }
 
 class Cliente extends Persona{
-    int credito;
+    double credito;
 }
 
 class Trabajador extends Persona{
-    int salario;
+    double salario;
 }
